Implementado a view de index de eventos de acordo com o prototipo.
