+++ Codigo/Core/Core.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="DTO\EventoIndexDTO.cs" />
     <Compile Remove="DTO\MaterialEstudoIndexDTO.cs" />
     <Compile Remove="Service\MaterialEstudoDTO.cs" />
   </ItemGroup>
