+++ Codigo/Core/DTO/EventoDTO.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
+﻿using System.ComponentModel;
 namespace Core.DTO
 {
     public class EventoDTO
@@ -11,5 +6,18 @@ namespace Core.DTO
         public int Id { get; set; }
         public string? Local { get; set; }
         public DateTime DataHoraInicio { get; set; }
+    }
+
+    public class EventoIndexDTO
+    {
+        public int Id { get; set; }
+        [DisplayName("Data hora início")]
+        public DateTime DataHoraInicio { get; set; }
+        [DisplayName("Local")]
+        public string? Local { get; set; }
+        [DisplayName("Planejados")]
+        public int Planejados { get; set; }
+        [DisplayName("Confirmados")]
+        public int Confirmados { get; set; }
     }
 }
