+++ Codigo/Service/EventoService.cs
@@ -84,7 +84,8 @@ namespace Service
                     Id = g.Id,
                     DataHoraInicio = g.DataHoraInicio,
                     Local = g.Local,
-                    Repertorio = g.Repertorio
+                    Planejados = 0,
+                    Confirmados = 0
                 }
                 ).AsNoTracking();
             return query;
@@ -99,7 +100,8 @@ namespace Service
                     Id = g.Id,
                     DataHoraInicio = g.DataHoraInicio,
                     Local = g.Local,
-                    Repertorio = g.Repertorio
+                    Planejados = 0,
+                    Confirmados = 0,
                 }
                 ).AsNoTracking();
             return query;
@@ -114,8 +116,7 @@ namespace Service
             if (request.Search != null && request.Search.GetValueOrDefault("value") != null)
             {
                 eventos = eventos.Where(g => g.DataHoraInicio.ToString().Contains(request.Search.GetValueOrDefault("value")!)
-                                                           || g.Local.ToString().Contains(request.Search.GetValueOrDefault("value")!)
-                                                           || g.Repertorio.ToString().Contains(request.Search.GetValueOrDefault("value")!));
+                                                           || g.Local.ToString().Contains(request.Search.GetValueOrDefault("value")!));
             }
 
             if (request.Order != null && request.Order[0].GetValueOrDefault("column")!.Equals("0"))
@@ -128,16 +129,9 @@ namespace Service
             else if (request.Order != null && request.Order[0].GetValueOrDefault("column")!.Equals("1"))
             {
                 if (request.Order[0].GetValueOrDefault("dir")!.Equals("asc"))
-                    eventos = eventos.OrderBy(g => g.DataHoraInicio);
+                    eventos = eventos.OrderBy(g => g.Local);
                 else
-                    eventos = eventos.OrderByDescending(g => g.DataHoraInicio);
-            }
-            else if (request.Order != null && request.Order[0].GetValueOrDefault("column")!.Equals("2"))
-            {
-                if (request.Order[0].GetValueOrDefault("dir")!.Equals("asc"))
-                    eventos = eventos.OrderBy(g => g.DataHoraInicio);
-                else
-                    eventos = eventos.OrderByDescending(g => g.DataHoraInicio);
+                    eventos = eventos.OrderByDescending(g => g.Local);
             }
 
             int countRecordsFiltered = eventos.Count();
